feat: change structure files and remove unnecessary components
