--- src/components/details/components/current-news/CurrentNews.tsx
@@ -1,31 +1,0 @@
-import { useRouter } from "next/router";
-import { Loader } from "@/components/common/loader/Loader";
-import { useArticles } from "@/api/articles/queries";
-import { hasSlugStringType } from "@/utils/type-guards";
-import { NewsDetails } from "../news-details/NewsDetails";
-
-export const CurrentNews: React.FC = () => {
-  const router = useRouter();
-  const titleFromUrl = router.query.id;
-  const { isPending, isError, data, error } = useArticles(
-    hasSlugStringType(titleFromUrl) ? titleFromUrl : ""
-  );
-
-  if (isPending) {
-    return <Loader />;
-  }
-
-  if (isError) {
-    return <span>Error: {error.message}</span>;
-  }
-
-  return (
-    <>
-      {data ? (
-        <NewsDetails article={data} />
-      ) : (
-        <p>{`We don't have any news on this request.`}</p>
-      )}
-    </>
-  );
-};
