--- src/components/home/Home.tsx
@@ -1,6 +1,0 @@
-import { PropsWithSlug } from "@/types";
-import { NewsList } from "../common/news-list/NewsList";
-
-export const Home: React.FC<PropsWithSlug> = ({ section }) => {
-  return <NewsList section={section} />;
-};
