--- src/pages/[section]/details/[id].tsx
@@ -1,5 +1,0 @@
-import { Details } from "@/components/details/Details";
-
-export default function DetailsPage() {
-  return <Details />;
-}
