+++ src/pages/[section]/index.tsx
@@ -1,9 +1,15 @@
 import { useRouter } from "next/router";
-import { Home } from "@/components/home/Home";
+import { NextPage } from "next";
+import { NewsList } from "@/components/news-list/NewsList";
+import { hasSlugStringType } from "@/utils/type-guards";
 
-export default function SectionPage() {
+const SectionPage: NextPage = () => {
   const router = useRouter();
+  const sectionName = hasSlugStringType(router.query.section)
+    ? router.query.section
+    : "";
 
-  
-return <Home section={router.query.section} />;
-}
+  return <NewsList section={sectionName} />;
+};
+
+export default SectionPage;
