--- src/types/index.ts
@@ -1,3 +1,0 @@
-export type PropsWithSlug = {
-  section: string | string[] | undefined;
-};
